WIP move performance tests to a dedicated folder
