+++ package.json
@@ -70,7 +70,7 @@
     "test:integration": "jest --runInBand --config=./test/integration/jest.config.js",
     "test:integration:coverage": "npm run test:integration -- --coverage",
     "test:unit:watch": "npm run test:unit:coverage -- --watchAll",
-    "test:e2e": "cross-env DEBUG=test JEST_IMAGE_SNAPSHOT_TRACK_OBSOLETE=1 JEST_PUPPETEER_CONFIG=./test/e2e/jest-puppeteer.config.js jest --runInBand --detectOpenHandles --testPathIgnorePatterns ./test/e2e/performance --config=./test/e2e/jest.config.js",
+    "test:e2e": "cross-env DEBUG=test JEST_IMAGE_SNAPSHOT_TRACK_OBSOLETE=1 JEST_PUPPETEER_CONFIG=./test/e2e/jest-puppeteer.config.js jest --runInBand --detectOpenHandles --testPathIgnorePatterns test/performance --config=./test/e2e/jest.config.js",
     "test:e2e:coverage": "npm run test:e2e -- --coverage",
     "test:perf": "cd ./test/e2e && jest --runInBand --detectOpenHandles --testNamePattern=performance",
     "test:bundles": "jest --runInBand --detectOpenHandles --config=./test/bundles/jest.config.js",
