+++ test/performance/bpmn.navigation.performance.test.ts
@@ -13,31 +13,51 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { getSimplePlatformName } from '../helpers/test-utils';
+import { delay, getSimplePlatformName } from '../e2e/helpers/test-utils';
+import { calculateMetrics, ChartData, PerformanceMetric } from '../e2e/helpers/perf-utils';
 import * as fs from 'fs';
-import { calculateMetrics, ChartData } from '../helpers/perf-utils';
-import { PageTester } from '../helpers/visu/PageTester';
-
-interface PerformanceMetric {
-  run: number;
-  TaskDuration: number;
-  ScriptDuration: number;
-  RecalcStyleDuration: number;
-  LayoutDuration: number;
-}
+import { PageTester } from '../e2e/helpers/visu/PageTester';
 
 const platform = getSimplePlatformName();
 const performanceDataFilePath = './performance/data/' + platform + '/data.js';
 const metricsArray: Array<PerformanceMetric> = [];
 
-describe.each([1, 2, 3, 4, 5])('load performance', run => {
+describe.each([1, 2, 3, 4, 5])('zoom performance', run => {
   // to have mouse pointer visible during headless test - add 'showMousePointer: true' as parameter
   const pageTester = new PageTester({ pageFileName: 'rendering-diagram', expectedPageTitle: 'BPMN Visualization - Diagram Rendering' });
-  const fileName = 'B.2.0';
 
-  it.each([1])('check performance for file loading and displaying diagram with FitType.HorizontalVertical', async () => {
+  const fileName = 'B.2.0';
+  let containerCenterX: number;
+  let containerCenterY: number;
+
+  beforeEach(async () => {
+    const bpmnContainerElementHandle = await pageTester.loadBPMNDiagramInRefreshedPage(fileName);
+    const bounding_box = await bpmnContainerElementHandle.boundingBox();
+    containerCenterX = bounding_box.x + bounding_box.width / 2;
+    containerCenterY = bounding_box.y + bounding_box.height / 2;
+  });
+
+  it.each([30])(`ctrl + mouse: check performance while performing zoom in and zoom out [%s times]`, async (xTimes: number) => {
+    const deltaX = -100;
     const metricsStart = await page.metrics();
-    await pageTester.loadBPMNDiagramInRefreshedPage(fileName);
+
+    // simulate mouse+ctrl zoom
+    await page.mouse.move(containerCenterX + 200, containerCenterY);
+    await page.keyboard.down('Control');
+    for (let i = 0; i < xTimes; i++) {
+      await page.mouse.wheel({ deltaX: deltaX });
+      if (i % 5 === 0) {
+        await delay(30);
+      }
+    }
+    await delay(100);
+    for (let i = 0; i < xTimes; i++) {
+      await page.mouse.wheel({ deltaX: -deltaX });
+      if (i % 5 === 0) {
+        await delay(30);
+      }
+    }
+    await delay(100);
     const metricsEnd = await page.metrics();
 
     const metric = { ...calculateMetrics(metricsStart, metricsEnd), run: run };
@@ -50,8 +70,8 @@ afterAll(() => {
     const oldDataString = fs.readFileSync(performanceDataFilePath, 'utf8');
     const oldData = JSON.parse(oldDataString.substring('const data = '.length, oldDataString.length)) as ChartData;
     const data = {
-      zoom: oldData.zoom,
-      load: oldData.load.concat(metricsArray),
+      zoom: oldData.zoom.concat(metricsArray),
+      load: oldData.load,
     };
     fs.writeFileSync(performanceDataFilePath, 'const data = ' + JSON.stringify(data));
   } catch (err) {
